Add a way to activate a virtualenv inside a project
In many python projects, the virtualenv is created at '.venv' inside
the project root. The virtualenvwrapper package has support for
automatically such virtualenvs by configuring the
'projectile-switch-project-action' variable, however it doesn't seem
to work with 'counsel-projectile'.

On the other hand 'venv-workon' only works on virtualenvs created
inside the 'venv-location' dir. So there's no way to activate a
virtualenv that's created inside the project.

This commit adds an interactive function that acts as a wrapper over
'venv'workon'. It checks if a dir '.venv' exists in the dir returned
by 'projectile-project-root'. In that case, it activates it. If not,
it falls back to calling 'venv-workon' normally.

This new function is bound to 'C-c v' in the python-mode-map.