Fix package archives and remove bootstrapping of use-package
In newer versions of emacs, bootstrapping of the 'use-package' package
redundant and hence been removed.

The package-archives are added using 'add-to-list' instead of 'setq'
so that the default values are also retained.

The org package archive at 'orgmode.org/elpa' is a broken
link (responds with 403).

This considerably reduces the time it takes for emacs to startup.