Prevent new eglot connection for crates in cargo registry
When 'M-.' (jump to function definition) opens a file from a crate
that's an external dependency (source files located inside
~/.cargo/registry), eglot ended up creating a new connection for each
external crate. This additional connection is not useful. Moreover
eglot connections don't get cleaned up even after all files in the
project are closed. This would result in a number of eglot
connections (and equivalent no. of rust-analyzer process) to be
running.

To avoid this, the 'eglot-ensure' call is being omitted for buffers of
rust src files from external crates.

When implementing this change, it was also realized that 'rustic-mode'
also configures lsp/eglot by default i.e. the 'eglot-ensure' added to
rust-mode-hook was not required. But now that more control is required
to create the connection selectively, rustic mode is configured to
skip lsp setup.