Fix coding system used when saving/writing files
Earlier, the the 2 variables 'buffer-file-coding-system' and
'coding-system-for-write' were being set. These caused the files to be
also written with unix style line endings (LF) regardless of the line
endings that emacs detected when reading the file.

This is particularly a problem when working on a code base that has
some files in the LF format and some in CRLF format, because always
saves the file in LF format.

The fix is to not set the above variables, so that while writing the
file back to the disk, emacs retains the same formatting that it
identified while reading it.